requirements.txt
aiodns==3.0.0
aiofile==3.5.0
aiohttp==3.7.4.post0
asgiref==3.3.4
async-timeout==3.0.1
attrs==21.2.0
caio==0.8.0
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
cryptography==3.4.7
Django==3.2.1
django-cors-headers==3.7.0
django-rest-knox==4.1.0
djangorestframework==3.12.4
idna==2.10
multidict==5.1.0
pkg-resources==0.0.0
pycares==4.0.0
pycparser==2.20
pycryptodomex==3.10.1
PyJWT==2.1.0
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.4
yarl==1.6.3
